Converted the activity to a fragment and tested it
