+++ app/src/main/AndroidManifest.xml
@@ -33,7 +33,7 @@
 
         <activity
             android:name=".activities.MainActivity"
-            android:exported="false" />
+            android:exported="true" />
         <activity
             android:name=".activities.DatabaseActivity"
             android:exported="false">
@@ -45,13 +45,6 @@
 
         <activity
             android:name=".activities.SharePostActivity"
-            android:exported="false">
-            <meta-data
-                android:name="android.app.lib_name"
-                android:value="" />
-        </activity>
-        <activity
-            android:name=".activities.AddNewPostActivity"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
